Fix Fasttext hash function for ngrams containing non-ASCII data
std::string in C++ uses signed char on most implementations. While the behavior
is implementation defined and binary Fasttext models trained after compiling
Fasttext with different compilers may result in different behavior, let's match
the behavior of the officially distributed binary models here.

+++ gluonnlp/vocab.py
@@ -628,7 +628,7 @@ class NGramHashes(SubwordFunction):
         s = s.encode(encoding)
         old_settings = np.seterr(all='ignore')
         for c in bytearray(s):
-            h = h ^ np.uint32(c)
+            h = h ^ np.uint32(np.int8(c))
             h = h * np.uint32(16777619)
         np.seterr(**old_settings)
         return h
